src/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.8.30
python-dotenv==1.0.1
python-telegram-bot==20.0
pytz==2024.2
six==1.16.0
tornado==6.4.1
tzlocal==5.2
